refactor: split non-schemaForm Builder code into __deprecated code path (#16567)
## What
We are approaching the end of the cycle, and the migration of Builder UI to SchemaForm still needs some more time before we fully switch all users over to it (still need to migrate the implementations for Jinja editors, AI-assist buttons, auto-created user inputs, and OAuth button UI to SchemaForm).

I chose not to add conditional logic based on the schema-form feature flag throughout my changes in this PR stack, because the switch to SchemaForm touches the state that sits at core of every part of the Builder form UI, and it would have slowed down development too much to maintain both sets of logic at the same time side-by-side. However, we still want to be able to merge these changes to master now, and allow users to select which version to use.

Therefore, @bnchrch and I have decided the best course of action is to duplicate the existing Builder implementation to a separate code path, and expose a toggle in the UI that allows users to switch between the old implementation and the new SchemaForm-based implementation.

This will allow the current stack of SchemaForm-Builder PRs to be merged to master and allow internal and interested users to start gaining value from the current state of the implementation.

## How
To produce this PR, copied the following folders into sibling folders with the suffix `__deprecated`:
- `oss/airbyte-webapp/src/components/connectorBuilder`
- `oss/airbyte-webapp/src/pages/connectorBuilder`
- `oss/airbyte-webapp/src/services/connectorBuilder`

which comprise the entirety of the Builder UI implementation.

**NOTE**: This is a temporary setup, and the `__deprecated` folders will be deleted once we feel ready to fully switch all users over to the SchemaForm-based Builder implementation.

I added the following logic to `oss/airbyte-webapp/src/pages/connectorBuilder/ConnectorBuilderRoutes.tsx`:
- If the new `"airbyte_connector-builder-advanced-mode"` local storage flag is false/unset, OR if the `"connectorBuilder.schemaForm"` experiment flag is not enabled, serve the `__deprecated` code path as the routes for the Connector Builder
- If both are enabled, then serve the non-__deprecated code path, which is the path that this stack of PRs is migrating to SchemaForm.

I [added](https://github.com/airbytehq/airbyte-platform-internal/pull/16567/files#diff-a81a5e8db6c0dc2058097fd3a0e0fbd9a60ab3fab296c0346a6e0f310c0334b1R49-R61) a Switch to the Sidebar in both the `__deprecated` and non-__deprecated Sidebar
- This is only rendered if the `"connectorBuilder.schemaForm"` experiment flag is enabled
- If the user toggles this on, then the `"airbyte_connector-builder-advanced-mode"` local storage flag is set to true and the page is refreshed so that the non-__deprecated code path with the SchemaForm-Builder implementation is served to the user

Here is a short recording demonstrating this in action:

https://github.com/user-attachments/assets/4e2765fe-0dc1-42b4-b497-1e2f268c544f

I also went through the code in the `__deprecated` code path to remove the few places that look at the `"connectorBuilder.schemaForm"` experiment flag, so that it guarantees we use the non-SchemaForm implementation when in that code path.

Pretty much every other change in this PR is just the copying of files to the `__deprecated` code paths, and deleting files that are unused in both code paths.