refactor: switch Builder sidebar to SchemaForm (#16312)
## What
This is the second PR in a stack of PRs to switch the Builder UI over to using SchemaForm under the hood.

> **Note on testing**: 
> - If running locally, I recommend running the [PR at the top of the stack](https://github.com/airbytehq/airbyte-platform-internal/pull/16452) to try things out, since not everything is fully working in these intermediate PRs. I split them up this way to make reviewing more approachable.
> - This also means the PR builds may not fully pass on these intermediate PRs, which is expected. I will squash everything before merging
> - You'll need to add `"connectorBuilder.schemaForm": true` to your `.experiments.json` file to try this out fully

This PR focuses on switching the base `BuilderSidebar` over to SchemaForm.

The theme of these PRs is to stop pulling values from `formValues` (which is the old UI state) and instead start pointing things to the `manifest` field in RHF, which is set by SchemaForm.

This PR also replaces the AddStream modal with a simple button. I may need to go back to the modal approach when I add back AI Assist functionality, but this keeps the approach and UX simpler for now.