refactor: builder schema form performance improvement (#16525)
## What
I noticed that when there are a lot of streams in the Builder, the performance was _very_ slow to the point that it felt almost unusable.

To see this, try importing [this manifest with a lot of streams](https://gist.github.com/lmossman/b5660c2359d95e8942dffcc6c8cf239a) into the base branch of this PR (`05-20-builder_schema_form_dynamic_streams`), and modify some of the inputs. You should see that it can take around a second or more per character for the UI to update.

## How
To solve this, I used the react profiler in browser dev tools to figure out what was causing the rendering to be so slow, and I methodically went through each slow area and corrected it to reduce the amount of components that were unnecessarily re-rendering.

One recurring issue is that pulling errors from `useFormState` causes a re-render every time the form values change, not just when the errors change. So I had to work around this a lot.

Here is an outline of the changes I made:
- New hashes of every stream were being calculated on every change to the form in `useStreamTestMetadata`. The `isStale` logic has been changed so that it is [set to true whenever a stream is modified](https://github.com/airbytehq/airbyte-platform-internal/pull/16525/files#diff-1faba83c23b37111daf13d994a835dd318844a7ff93fd6596daf8e9490130070R165), rather than calculating the hash and comparing it to the old one on the fly to determine if it is stale.
- `BuilderSidebar` was refactored to call `getValues` to fetch stream information and add a custom hook `useStreamsSidebarData` that extracts only the minimal info needed for the sidebar, so that the whole sidebar stops re-rendering on every change to the form
- `permission` was pulled out of the Builder form state and into its own hook, as this was causing any component that used `permission` to be re-rendered on every form change
  - A lot of the updates to files were caused by this change since a lot of components use `permissions`
- `StreamConfigView` was refactored to only `watch` the paths that it actually needs, rather than the entire stream, and the `hasErrors` logic was moved into the tab components where it was actually needed, so that the entire stream view doesn't re-render when the errors change
- `PublishButton` was refactored to pull just the info that is actually needed from hooks now that we have moved from `formValues` to `manifest`
- `ObjectControl` was changed to only pull errors where needed so that the entire object and all of its children don't get re-rendered any time there is a form change. This means we no longer show the error indicator on Advanced collapsibles but the performance gain is worth it.
- `DynamicValidator` was refactored to only process the paths that change when form values are modified (and their sub-paths), rather than process every path in the form on any change
- `getSchemaAtPath` in `SchemaForm` was corrected to not throw an error when no option is selected
- `useErrorAtPath` was optimized to use `getFieldState` instead of `useFormState` to fetch errors for a particular path, so that it doesn't re-render any time any value in the form is changed

With these changes, if you try again with the same [manifest with a lot of streams](https://gist.github.com/lmossman/b5660c2359d95e8942dffcc6c8cf239a), you should notice a vast performance improvement.