chore: Remove errant Workspace#list entity operation
The expected entity operations for downstream tooling are one of `create`, `read`, `update`, and `delete`:

```
Error: Unexpected op annotation list applied as Workspace#list: expected it to be one of ["create","read","update","delete"]
```

Since the `listWorkspaces` API operation appears to be intended as its own Terraform data resource and its usage would be separate from the workspace managed resource, the entity operation could be updated to create a workspaces data resource via `Workspaces#read`. However, it is removed for now to match the current behavior of not doing anything in the Terraform code generation while also preventing the error.