fix: multi option selection input type change (#16983)
## What
There is a small bug in SchemaForm, where if there is a field which is a `oneOf` with the first option having `type: integer` and the second option having `type: string`, and if the current value of the input is `null` (meaning the `integer` option is selected), then if you select the `string` option from the dropdown, nothing happens.

This is happening because when `setValue` in the `onSelect` callback of the MultiOptionControl listbox is triggered, the numeric InputWrapper's `setValueAs` logic causes `null` to be set instead of the empty string `""`.

Here is a recording demonstrating this issue with the `page_size` field in the paginator component of the Builder:

https://github.com/user-attachments/assets/7e4ebc70-1e09-41d3-bad6-d0dfdce5cd3d

## How
The workaround for this issue is to unregister the field before calling `setValue`, as this will allow the value to be set without passing through `setValueAs`, and once the value is set, the SchemaFormControl is re-rendered anyway, causing the field to be re-registered with the new type.

Here is a recording demonstrating the fix:

https://github.com/user-attachments/assets/a594c37f-14f4-4ab4-bc3b-e4a61632dae2

I also added a check that if the newly selected option is the same as the currently selected option, then it just returns from `onSelect`, since nothing needs to happen in that case. This prevents selecting the same option from clearing out the values.