chore: centralizing where we set the log level (#17333)
Centralizing where we set the log level to ensure it is set in the orchestrator at the same level as the destination and the source.

This PR updates where this is all set from the EnvVarConfigBeanFactory.kt to the RuntimeEnvVarFactory.kt.
That allows us to avoid failure due to the env variable being added again.