fix: builder stream test button validation (#17335)
## What
There are a couple of small issues with the Builder validation UX currently:
1. If you create a brand new builder connector without AI assist, then click the `Test` button without making any changes, it tries to run a test read and fails, whereas it should be highlighting the stream name input with a `Required` error state and focusing that input to indicate to the user it should be filled out first
2. With a valid stream open, if you enable Incremental Sync and then click the `Test` button, it scrolls to and highlights the Cursor field with a `Required` error state, but doesn't focus the field

## How
For issue 1, the problem is that when we call `trigger()` in `useBuilderErrors`, in the `then` handler, the `errors` object from `useFormState` hasn't had time to get the new errors value, because validation was _just_ triggered on that render cycle.

Rather than manually waiting for the next render cycle, I solved this instead by explicitly getting the manifest error state using `getFieldState`, and then passing the result to the later functions for processing. This way, the up-to-date errors are retrieved when they are needed.

---

For issue 2, the problem is that if the field being focused supports jinja, then it is actually a monaco editor rather than a normal input (so that we can render jinja suggestions and highlighting). 

So, the current `useFocusField` logic was not able to focus it because the element with the corresponding `data-field-path` prop is not actually an input. To fix this, I simply added a case that looks for the `textarea` child of the element instead, which the monaco editor does have, and this allows it to be properly focused.