chore: skip Source check if running clear. (#17585)
It's not clear if this was ever working. The current code has a check on the connectionUpdaterInput.resetConnection field, however this field is never set and defaults to false. This means we always run Source checks on reset. This is unnecessary since Clear reads from an empty source.

This PR fixes this.

Some of this PR is verbose due to Temporal versioning. Strategy: push out a version, give a few days for daily syncs to move onto the new workflow, and then delete this. Goal here is to minimise non-determinism errors from workflow restarts in the workflow restarter cron.

The existing field is leaked into workflow runs. This info is only needed when deciding to run check. We want to delete this field, which means not using the existing field.

Instead add two methods to the job status activity shouldRunSourceCheck and shouldRunDestinationCheck. The workflow directly calls these in order to understand whether check should run. These methods can be used by the next iteration of the orchestration workflow to do the same, so there is slight future proofing here.